move canvas init to own component + new store variable
