+++ src/store/map.ts
@@ -1,11 +1,11 @@
 import { writable, type Writable } from 'svelte/store';
 import type { Map as MMap } from 'maplibre-gl';
 import MapboxDraw from "@mapbox/mapbox-gl-draw"
-import { CUSTOM_GL_DRAW_STYLES } from '../lib/gl_draw_styles.js'
-import { PolygonFourPointsOnly } from '../lib/custom_poly.js'
-import { DeleteClickedZone } from '../lib/custom_delete.js'
+import { CUSTOM_GL_DRAW_STYLES } from '$lib/gl_draw_styles.js'
+import { PolygonFourPointsOnly } from '$lib/custom_poly.js'
+import { DeleteClickedZone } from '$lib/custom_delete.js'
 
-export const map: Writable<MMap> = writable();
+export const map: Writable<MMap> = writable()
 export const draw: Writable<MapboxDraw> = writable(new MapboxDraw({
     userProperties: true,
     displayControlsDefault: false,
